dedup GetPassthroughAnnotations
